Move LLObject.h to DataStructure
