+++ llama_cpp/example/llobject/llobject.cpp
@@ -15,7 +15,7 @@
 //  limitations under the License.
 //
 
-#include <Utility/LLObject.h>
+#include <DataStructure/LLObject.h>
 #include <iostream>
 #include <string>
